refac: move get_weather_scenarios to _helpers.py
