+++ scripts/_helpers.py
@@ -50,6 +50,15 @@ ENERGY_UNITS = {"TWh", "GWh", "MWh", "kWh"}
 POWER_UNITS = {"GW", "MW", "kW"}
 PRICE_UNITS = {"EUR/MWh", "EUR/MWh_e", "EUR/MWh_H2"}
 
+# Weather scenarios that contain data in the TYNDP 2026 data,
+# per planning horizon.
+AVAILABLE_WEATHER_SCENARIOS = {
+    2030: [3, 21, 29],
+    2035: [32, 37, 59],
+    2040: [65, 71, 77],
+    2050: [91, 92, 106],
+}
+
 PYPSA_V1 = bool(re.match(r"^1\.\d", pypsa.__version__))
 
 
@@ -1934,3 +1943,39 @@ def parse_weather_scenario(s: pd.Series) -> pd.Series:
     Convert weather scenario labels (eg. WS065) into their integer index.
     """
     return pd.to_numeric(s.astype(str).str.removeprefix("WS"), errors="coerce")
+
+
+def get_weather_scenario(weather_scenarios, pyear):
+    """
+    Select the weather scenario to use for a given planning year.
+
+    Parameters
+    ----------
+    weather_scenarios : dict
+        Mapping of planning year to a list of requested weather scenarios,
+        e.g. ``{pyear: [weather_scenario, ...]}``.
+    pyear : int
+        Planning year for which to select the weather scenario.
+
+    Returns
+    -------
+    int
+        Selected weather scenario. Falls back to the first entry in
+        ``AVAILABLE_WEATHER_SCENARIOS[pyear]`` if unavailable.
+
+    Notes
+    -----
+    Currently always picks the first requested weather scenario; should be
+    adapted once the full weather year implementation is available in SB.
+    """
+    weather_scenario = weather_scenarios[pyear][0]
+
+    if weather_scenario not in AVAILABLE_WEATHER_SCENARIOS[pyear]:
+        fallback_scenario = AVAILABLE_WEATHER_SCENARIOS[pyear][0]
+        logger.warning(
+            f"Weather scenario WS{weather_scenario:03d} not available for "
+            f"planning year {pyear}, falling back to WS{fallback_scenario:03d}"
+        )
+        weather_scenario = fallback_scenario
+
+    return weather_scenario
